refactor: move phase-6 components into layout/ category
Align new page layout components with the categorized folder structure
introduced in #27. Moves page, section, top-title-bar,
title-bar-title-group, and collection into layout/.
