+++ src/components/layout/top-title-bar/rr-top-title-bar.stories.js
@@ -2,7 +2,7 @@ import { html } from 'lit';
 import './rr-top-title-bar.ts';
 
 export default {
-  title: 'Components/Top Title Bar',
+  title: 'Components/Layout/Top Title Bar',
   component: 'rr-top-title-bar',
   tags: ['autodocs'],
   parameters: {
